+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.6.0
+
+- **BREAKING**: The `snapshot` signature has changed to make it more predictable
+where the snapshot gets written to. In most cases `snapshot(__filename)` will
+suffice.
+
 ## 0.5.0
 
 - initial pre-release